requirements.txt
Django~=4.1.3
django-environ~=0.9.0
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.2.2
Faker~=15.3.4
django-cors-headers==3.13.0
django-filter==22.1
mysqlclient==2.1.1